Add bans to the Expanded format
Not sure if this is all that needs done in order to add these cards to the expanded banlist, but following the [official Pokemon TCG banlist](https://www.pokemon.com/us/pokemon-tcg-banned-card-list/), these need added. If they are needed elsewhere, please let me know where to add the commits necessary. Here's the numbers as they appear in the list to help make it easier see what was added.

361-124 / 371-71 - Red Card (XY / Generations)
410-SM85 / 414-45 - Marshadow (Sun & Moon Promo / Shining Legends)
417-83 - Flabebe (Forbidden Light)
420-91 - Unown (Lost Thunder)
423-78 - Mismagius (Unbroken Bonds)
423-165 - Chip-Chip Ice Axe (Unbroken Bonds)
423-178 / 425-60 - Lt. Surge's Strategy (Unbroken Bonds / Hidden Fates)
424-206 / 424-253 - Reset Stamp (Unified Minds)
425-58 / 425-68 - Jessie & James (Hidden Fates)
426-194 / 426-265 - Island Challenge Amulet (Cosmic Eclipse)