Updated the Expanded format bans
Not sure if this is all that needs done in order to add these cards to the expanded banlist, but following the [official Pokemon TCG banlist](https://www.pokemon.com/us/pokemon-tcg-banned-card-list/), these need added. Please let me know if anything is missing. Here's the numbers as they appear in the list to help make it easier see what was added.

361-124 / 371-71 - **Red Card** _(XY / Generations)_
410-SM85 / 414-45 - **Marshadow** _(Sun & Moon Promo / Shining Legends)_
417-83 - **Flabebe** _(Forbidden Light)_
420-91 - **Unown** _(Lost Thunder)_
423-78 - **Mismagius** _(Unbroken Bonds)_
423-165 - **Chip-Chip Ice Axe** _(Unbroken Bonds)_
423-178 / 425-60 - **Lt. Surge's Strategy** _(Unbroken Bonds / Hidden Fates)_
424-206 / 424-253 - **Reset Stamp** _(Unified Minds)_
425-58 / 425-68 - **Jessie & James** _(Hidden Fates)_
427-194 / 427-265 - **Island Challenge Amulet** _(Cosmic Eclipse)_