Re-order and add missing Sun and Moon promos
I can add the scans that may be missing, although I am aware we need to re-structure our scans on the server first before that is done. Regardless, I went ahead and added all the missing stuff. There are still some promos from the latest box that was just released that need added, but I still need to look into any promos that would need implemented before doing all of that. Not to mention, I'll have to make the db entries for the new box myself since neither pokemontcg.io or Kirby's repo has the new promos on their files.