Read expansion data prior to reading card data
When using the pio v2 data, the expansions are now in their own JSON file.
These changes read that data before reading card data so that the relevant
expansion data can be used to generate files. This should reduce the number of
manual changes that need to be made to YAML and groovy files, and fixes
downloading scans using pio instead of kirby. Also fixes the issue where using
multiple pios would throw everything in the same files because of the naming
issue.