Add files via upload
The YAML contains all the meta information of the cards.

The other TXT document is a helpful clarifying aid to disambiguate which Promo is which. It tells what Promotion it is from, for example CoroCoro Comic April 1998 issue insert.  Also included in it are the bulbapedia links to the cards/attacks/artwork themselves